mysql/requirements.txt
# Core dependencies
mcp[cli]>=0.1.0
pymysql>=1.1.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0

# HTTP server dependencies
uvicorn>=0.23.0
fastapi>=0.103.0
httpx>=0.24.1

# Development dependencies (uncomment as needed)
# pytest>=7.0.0
# black>=23.0.0
# isort>=5.12.0
# flake8>=6.0.0
# mypy>=1.0.0
# pytest-cov>=4.0.0 

# Development dependencies
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
pytest-cov>=4.0.0 